refactor(catalog-policy): move IMediaCatalogEvaluationRepository to domain layer
Move repository interface from infrastructure to domain layer for Clean
Architecture compliance. Application layer now imports from domain/repositories
instead of infrastructure/repositories.

- Create domain/repositories/media-catalog-evaluation.repository.interface.ts
- Update domain/repositories/index.ts with new export
- Add re-export in infrastructure for backward compatibility
- Update all consumers to import from domain layer
