+++ apps/api/src/modules/catalog-policy/application/services/catalog-evaluation.service.spec.ts
@@ -13,9 +13,11 @@ import {
   EvaluationReason,
   type EvaluationContextType,
 } from '../../domain/constants/evaluation.constants';
-import { POLICY_INPUT_REPOSITORY } from '../../domain/repositories/policy-input.repository.interface';
+import {
+  POLICY_INPUT_REPOSITORY,
+  MEDIA_CATALOG_EVALUATION_REPOSITORY,
+} from '../../domain/repositories';
 import type { PolicyEngineInput, PolicyConfig } from '../../domain/types/policy.types';
-import { MEDIA_CATALOG_EVALUATION_REPOSITORY } from '../../infrastructure/repositories/media-catalog-evaluation.repository';
 
 import { CatalogEvaluationService } from './catalog-evaluation.service';
 import { CatalogPolicyService } from './catalog-policy.service';
