+++ apps/api/src/modules/catalog-policy/domain/repositories/index.ts
@@ -9,3 +9,8 @@ export {
   type IPolicyInputRepository,
   type FetchBatchIdsOptions,
 } from './policy-input.repository.interface';
+
+export {
+  MEDIA_CATALOG_EVALUATION_REPOSITORY,
+  type IMediaCatalogEvaluationRepository,
+} from './media-catalog-evaluation.repository.interface';
